# onboarding-common/src/main/java/org/acumos/onboarding/services/impl/OnboardingController.java
/*-
 * ===============LICENSE_START=======================================================
 * Acumos
 * ===================================================================================
 * Copyright (C) 2017 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
 * ===================================================================================
 * This Acumos software file is distributed by AT&T and Tech Mahindra
 * under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * This file is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * ===============LICENSE_END=========================================================
 */

package org.acumos.onboarding.services.impl;

import java.io.File;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.UUID;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.ws.rs.Consumes;
import javax.ws.rs.core.MediaType;

import org.acumos.cds.CodeNameType;
import org.acumos.cds.domain.MLPCodeNamePair;
import org.acumos.cds.domain.MLPSolution;
import org.acumos.cds.domain.MLPSolutionRevision;
import org.acumos.cds.domain.MLPUser;
import org.acumos.cds.transport.AuthorTransport;
import org.acumos.cds.transport.RestPageRequest;
import org.acumos.cds.transport.RestPageResponse;
import org.acumos.onboarding.common.exception.AcumosServiceException;
import org.acumos.onboarding.common.models.OnboardingNotification;
import org.acumos.onboarding.common.models.ServiceResponse;
import org.acumos.onboarding.common.utils.Crediantials;
import org.acumos.onboarding.common.utils.EELFLoggerDelegate;
import org.acumos.onboarding.common.utils.JsonRequest;
import org.acumos.onboarding.common.utils.LogBean;
import org.acumos.onboarding.common.utils.LogThreadLocal;
import org.acumos.onboarding.common.utils.OnboardingConstants;
import org.acumos.onboarding.common.utils.UtilityFunction;
import org.acumos.onboarding.component.docker.preparation.Metadata;
import org.acumos.onboarding.component.docker.preparation.MetadataParser;
import org.acumos.onboarding.logging.OnboardingLogConstants;
import org.acumos.onboarding.services.DockerService;
import org.json.simple.JSONObject;
import org.slf4j.MDC;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.client.HttpClientErrorException;
import org.springframework.web.multipart.MultipartFile;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@RestController
@RequestMapping(value = "/v2")
@Api(value = "Operation to to onboard a ML model", tags = "Onboarding Service APIs")
/**
 *
 * @author *****
 *
 */
public class OnboardingController extends CommonOnboarding implements DockerService {
	private static EELFLoggerDelegate logger = EELFLoggerDelegate.getLogger(OnboardingController.class);
	Map<String, String> artifactsDetails = new HashMap<>();
	public static String lOG_DIR_LOC = "/maven/logs/on-boarding/applog";

	public OnboardingController() {
		// Property values are injected after the constructor finishes
	}

	@SuppressWarnings("unchecked")
	@Override
	@Consumes(MediaType.MULTIPART_FORM_DATA)
	@ApiOperation(value = "Check User authentication and returns JWT token", response = ServiceResponse.class)
	@ApiResponses(value = {
			@ApiResponse(code = 500, message = "Something bad happened", response = ServiceResponse.class),
			@ApiResponse(code = 400, message = "Invalid request", response = ServiceResponse.class)	})
	@RequestMapping(value = "/auth", method = RequestMethod.POST, produces = "application/json")
	public ResponseEntity<ServiceResponse> OnboardingWithAuthentication(@RequestBody JsonRequest<Crediantials> cred,
			HttpServletResponse response) throws AcumosServiceException {
		logger.debug(EELFLoggerDelegate.debugLogger, "Started User Authentication");
		try {
			Crediantials obj = cred.getBody();

			String user = obj.getUsername();
			String pass = obj.getPassword();

			JSONObject crediantials = new JSONObject();
			crediantials.put("username", user);
			crediantials.put("password", pass);

			JSONObject reqObj = new JSONObject();
			reqObj.put("request_body", crediantials);

			String token = portalClient.loginToAcumos(reqObj);

			if (token != null) {
				// Setting JWT token in header
				response.setHeader("jwtToken", token);
				logger.debug(EELFLoggerDelegate.debugLogger, "User Authentication Successful");
				return new ResponseEntity<ServiceResponse>(ServiceResponse.successJWTResponse(token), HttpStatus.OK);
			} else {
				logger.debug(EELFLoggerDelegate.debugLogger, "Either Username/Password is invalid.");
				throw new AcumosServiceException(AcumosServiceException.ErrorCode.OBJECT_NOT_FOUND,
						"Either Username/Password is invalid.");
			}

		} catch (AcumosServiceException e) {
			logger.error(EELFLoggerDelegate.errorLogger, e.getMessage(), e);
			return new ResponseEntity<ServiceResponse>(ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage()),
					HttpStatus.UNAUTHORIZED);
		}
	}

	/************************************************
	 * End of Authentication
	 *****************************************************/

	@Override
	@Consumes(MediaType.MULTIPART_FORM_DATA)
	@ApiOperation(value = "Upload model file and its meta data as string to dockerize", response = ServiceResponse.class)
	@ApiResponses(value = {
			@ApiResponse(code = 200, message = "Created", response = ServiceResponse.class),
			@ApiResponse(code = 500, message = "Something bad happened", response = ServiceResponse.class),
			@ApiResponse(code = 400, message = "Invalid request", response = ServiceResponse.class),
			@ApiResponse(code = 401, message = "Unauthorized User", response = ServiceResponse.class) })
	@RequestMapping(value = "/models", method = RequestMethod.POST, produces = "application/json")
	public ResponseEntity<ServiceResponse> onboardModel(HttpServletRequest request,
			@RequestPart(required = true) MultipartFile model, @RequestPart(required = true) MultipartFile metadata,
			@RequestPart(required = true) MultipartFile schema,
			@RequestHeader(value = "Authorization", required = false) String authorization,
			@RequestHeader(value = "tracking_id", required = false) String trackingID,
			@RequestHeader(value = "provider", required = false) String provider,
			@RequestHeader(value = "shareUserName", required = false) String shareUserName,
			@RequestHeader(value = "modName", required = false) String modName,
			@RequestHeader(value = "deployment_env", required = false) Integer deployment_env,
			@RequestHeader(value = "Request-ID", required = false) String request_id) throws AcumosServiceException {

		OnboardingNotification onboardingStatus = null;

		if (trackingID != null) {
			logger.debug(EELFLoggerDelegate.debugLogger, "Tracking ID: {}", trackingID);
		} else {
			trackingID = UUID.randomUUID().toString();
			logger.debug(EELFLoggerDelegate.debugLogger, "Tracking ID Created: {}", trackingID);
		}

		if (request_id != null) {
			logger.debug(EELFLoggerDelegate.debugLogger, "Request ID: {}", request_id);
		} else {
			request_id = UUID.randomUUID().toString();
			logger.debug(EELFLoggerDelegate.debugLogger, "Request ID Created: {}", request_id);
		}

		//code to retrieve the current pom version
		//UtilityFunction.getCurrentVersion();
		onboardingStatus = new OnboardingNotification(cmnDataSvcEndPoinURL, cmnDataSvcUser, cmnDataSvcPwd, request_id);
		onboardingStatus.setTrackingId(trackingID);
		onboardingStatus.setRequestId(request_id);
		MDC.put(OnboardingLogConstants.MDCs.REQUEST_ID, request_id);

		String fileName = "OnboardingLog.txt";
		LogBean logBean = new LogBean();
		logBean.setLogPath(lOG_DIR_LOC + File.separator + trackingID);
		logBean.setFileName(fileName);
		LogThreadLocal logThread = new LogThreadLocal();
		logThread.set(logBean);
		// create log file to capture logs as artifact
		UtilityFunction.createLogFile();

		String version = appVersion;
		logger.debug(EELFLoggerDelegate.debugLogger,"On-boarding version {}", version);

		MLPUser shareUser = null;
		Metadata mData = null;
		String modelName = null;

		try {
			// 'authorization' represents JWT token here...!
			if (authorization == null) {
				logger.error(EELFLoggerDelegate.errorLogger, "Token Not Available...!");
				throw new AcumosServiceException(AcumosServiceException.ErrorCode.OBJECT_NOT_FOUND,
						"Token Not Available...!");
			}

			if (shareUserName != null) {
				RestPageResponse<MLPUser> user = cdmsClient.findUsersBySearchTerm(shareUserName,
						new RestPageRequest(0, 9));

				List<MLPUser> uList = user.getContent();

				if (uList.isEmpty()) {
					logger.error(EELFLoggerDelegate.errorLogger,
							"User " + shareUserName + " not found: cannot share model; onboarding aborted");
					throw new AcumosServiceException(AcumosServiceException.ErrorCode.OBJECT_NOT_FOUND,
							"User " + shareUserName + " not found: cannot share model; onboarding aborted");
				} else {
					shareUser = uList.get(0);
				}
			}

			// Call to validate Token .....!
			String ownerId = validate(authorization, provider);

			if (ownerId != null && !ownerId.isEmpty()) {

				logger.debug(EELFLoggerDelegate.debugLogger, "Token validation successful");

				// update userId in onboardingStatus
				if (onboardingStatus != null)
					onboardingStatus.setUserId(ownerId);

				logger.debug(EELFLoggerDelegate.debugLogger,
						"Onboarding request recieved with " + model.getOriginalFilename());

				// Notify Create solution or get existing solution ID has
				// started
				if (onboardingStatus != null) {
					onboardingStatus.notifyOnboardingStatus("CreateSolution", "ST", "CreateSolution Started");
				}

				modelOriginalName = model.getOriginalFilename();
				String modelId = UtilityFunction.getGUID();
				File outputFolder = new File("tmp", modelId);
				outputFolder.mkdirs();
				boolean isSuccess = false;
				MLPSolution mlpSolution = null;
				File localmodelFile = new File(outputFolder, model.getOriginalFilename());
				File localMetadataFile = new File(outputFolder, metadata.getOriginalFilename());
				File localProtobufFile = new File(outputFolder, schema.getOriginalFilename());
				MLPSolutionRevision revision;

				try {

					try {
						// Notify Create solution or get existing solution ID
						// has
						// started
						if (onboardingStatus != null) {
							onboardingStatus.notifyOnboardingStatus("CreateSolution", "ST", "CreateSolution Started");
						}

						UtilityFunction.copyFile(model.getInputStream(), localmodelFile);

						UtilityFunction.copyFile(metadata.getInputStream(), localMetadataFile);

						UtilityFunction.copyFile(schema.getInputStream(), localProtobufFile);

						metadataParser = new MetadataParser(localMetadataFile);
						mData = metadataParser.getMetadata();

						mData.setOwnerId(ownerId);

						List<MLPSolution> solList = getExistingSolution(mData);

						boolean isListEmpty = solList.isEmpty();

						if (isListEmpty) {
							mlpSolution = createSolution(mData, onboardingStatus);
							mData.setSolutionId(mlpSolution.getSolutionId());
							logger.debug(EELFLoggerDelegate.debugLogger,
									"New solution created Successfully " + mlpSolution.getSolutionId());
						} else {
							logger.debug(EELFLoggerDelegate.debugLogger,
									"Existing solution found for model name " + solList.get(0).getName());
							mlpSolution = solList.get(0);
							mData.setSolutionId(mlpSolution.getSolutionId());
						}

						revision = createSolutionRevision(mData);

						modelName = mData.getModelName() + "_" + mData.getSolutionId();

						// Solution id creation completed
						// Notify Creation of solution ID is successful
						if (onboardingStatus != null) {
							// set solution Id
							if (mlpSolution.getSolutionId() != null) {
								onboardingStatus.setSolutionId(mlpSolution.getSolutionId());
							}
							// set revision id
							if (mData.getRevisionId() != null) {
								onboardingStatus.setRevisionId(mData.getRevisionId());
							}
							// notify
							onboardingStatus.notifyOnboardingStatus("CreateSolution", "SU",
									"CreateSolution Successful");
						}
					} catch (AcumosServiceException e) {
						HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
						logger.error(EELFLoggerDelegate.errorLogger, e.getErrorCode() + "  " + e.getMessage());
						if (e.getErrorCode().equalsIgnoreCase(OnboardingConstants.INVALID_PARAMETER)) {
							httpCode = HttpStatus.BAD_REQUEST;
						}
						// Create Solution failed. Notify
						if (onboardingStatus != null) {
							// notify
							onboardingStatus.notifyOnboardingStatus("CreateSolution", "FA", e.getMessage());
						}
						return new ResponseEntity<ServiceResponse>(
								ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage(), modelName), httpCode);
					} catch (Exception e) {
						logger.error(EELFLoggerDelegate.errorLogger, e.getMessage());
						// Create Solution failed. Notify
						if (onboardingStatus != null) {
							// notify
							onboardingStatus.notifyOnboardingStatus("CreateSolution", "FA", e.getMessage());
						}
						if (e instanceof AcumosServiceException) {
							return new ResponseEntity<ServiceResponse>(
									ServiceResponse.errorResponse(((AcumosServiceException) e).getErrorCode(),
											e.getMessage(), modelName),
									HttpStatus.INTERNAL_SERVER_ERROR);
						} else {
							return new ResponseEntity<ServiceResponse>(
									ServiceResponse.errorResponse(AcumosServiceException.ErrorCode.UNKNOWN.name(),
											e.getMessage(), modelName),
									HttpStatus.INTERNAL_SERVER_ERROR);
						}
					}

					String actualModelName = getActualModelName(mData, mlpSolution.getSolutionId());
					// Add artifacts started. Notification will be handed by
					// addArtifact method itself for started/success/failure
					artifactsDetails = getArtifactsDetails();

					addArtifact(mData, localmodelFile, getArtifactTypeCode("Model Image"), mData.getModelName(),
							onboardingStatus);

					addArtifact(mData, localProtobufFile, getArtifactTypeCode("Model Image"), mData.getModelName(),
							onboardingStatus);

					addArtifact(mData, localMetadataFile, getArtifactTypeCode("Metadata"), mData.getModelName(),
							onboardingStatus);

					// Notify TOSCA generation started
					if (onboardingStatus != null) {
						onboardingStatus.notifyOnboardingStatus("CreateTOSCA", "ST", "TOSCA Generation Started");
					}

					generateTOSCA(localProtobufFile, localMetadataFile, mData, onboardingStatus);

					// Notify TOSCA generation successful
					if (onboardingStatus != null) {
						onboardingStatus.notifyOnboardingStatus("CreateTOSCA", "SU", "TOSCA Generation Successful");
					}

					// call microservice
					logger.debug(EELFLoggerDelegate.debugLogger, "Before microservice call Parameters : SolutionId "
							+ mlpSolution.getSolutionId() + " and RevisionId " + revision.getRevisionId());
					try {
						ResponseEntity<ServiceResponse> response = microserviceClient.generateMicroservice(
								mlpSolution.getSolutionId(), revision.getRevisionId(), provider, authorization,
								trackingID, modName, deployment_env, request_id);
						if (response.getStatusCodeValue() == 200 || response.getStatusCodeValue() == 201) {
							isSuccess = true;
						}
					} catch (Exception e) {
						logger.error(EELFLoggerDelegate.errorLogger,
								"Exception occured while invoking microservice API " + e);
						throw e;
					}

					// Model Sharing
					if (isSuccess && (shareUserName != null) && revision.getRevisionId() != null) {
						try {
							AuthorTransport author = new AuthorTransport(shareUserName, shareUser.getEmail());
							AuthorTransport authors[] = new AuthorTransport[1];
							logger.debug(EELFLoggerDelegate.debugLogger,
									"Author Name " + author.getName() + " and Email " + author.getContact());
							authors[0] = author;
							revision.setAuthors(authors);
							cdmsClient.updateSolutionRevision(revision);
							logger.debug(EELFLoggerDelegate.debugLogger,
									"Model Shared Successfully with " + shareUserName);
						} catch (Exception e) {
							isSuccess = false;
							logger.error(EELFLoggerDelegate.errorLogger, " Failed to share Model", e);
							throw e;
						}
					}

					ResponseEntity<ServiceResponse> res = new ResponseEntity<ServiceResponse>(
							ServiceResponse.successResponse(mlpSolution), HttpStatus.CREATED);
					logger.debug(EELFLoggerDelegate.debugLogger,
							"Onboarding is successful for model name: " + mlpSolution.getName() + ", SolutionID: "
									+ mlpSolution.getSolutionId() + ", Status Code: " + res.getStatusCode());
					return res;
				} finally {

					try {
						UtilityFunction.deleteDirectory(outputFolder);
						if (isSuccess == false) {
							logger.debug(EELFLoggerDelegate.debugLogger,
									"Onboarding Failed, Reverting failed solutions and artifacts.");
							if (metadataParser != null && mData != null) {
								revertbackOnboarding(metadataParser.getMetadata(), mlpSolution.getSolutionId());
							}
						}

						// push docker build log into nexus

						File file = new java.io.File(lOG_DIR_LOC + File.separator + trackingID
								+ File.separator + fileName);
						logger.debug(EELFLoggerDelegate.debugLogger, "Log file length " + file.length(), file.getPath(),
								fileName);
						if (metadataParser != null && mData != null) {
							logger.debug(EELFLoggerDelegate.debugLogger,
									"Adding of log artifacts into nexus started " + fileName);

							// String nexusArtifactID = "onboardingLog_"+trackingID;
							String nexusArtifactID = "OnboardingLog";

							addArtifact(mData, file, getArtifactTypeCode(OnboardingConstants.ARTIFACT_TYPE_LOG),
									nexusArtifactID, onboardingStatus);
							MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
									OnboardingLogConstants.ResponseStatus.COMPLETED.name());
							logger.debug(EELFLoggerDelegate.debugLogger, "Artifacts log pushed to nexus successfully",
									fileName);
						}

						// delete log file
						UtilityFunction.deleteDirectory(file);
						logThread.unset();
						mData = null;
					} catch (AcumosServiceException e) {
						MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
								OnboardingLogConstants.ResponseStatus.ERROR.name());
						MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION,
								OnboardingLogConstants.ResponseStatus.ERROR.name());
						mData = null;
						logger.error(EELFLoggerDelegate.errorLogger, "RevertbackOnboarding Failed");
						HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
						return new ResponseEntity<ServiceResponse>(
								ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage(), modelName), httpCode);
					}
				}
			} else {
				try {
					MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
							OnboardingLogConstants.ResponseStatus.ERROR.name());
					MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, "Either Username/Password is invalid.");
					logger.error(EELFLoggerDelegate.errorLogger, "Either Username/Password is invalid.");
					throw new AcumosServiceException(AcumosServiceException.ErrorCode.INVALID_TOKEN,
							"Either Username/Password is invalid.");
				} catch (AcumosServiceException e) {
					return new ResponseEntity<ServiceResponse>(
							ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage()), HttpStatus.UNAUTHORIZED);
				}
			}

		} catch (AcumosServiceException e) {

			MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
					OnboardingLogConstants.ResponseStatus.ERROR.name());
			MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, e.getMessage());
			HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
			logger.error(EELFLoggerDelegate.errorLogger, e.getErrorCode() + "  " + e.getMessage());
			if (e.getErrorCode().equalsIgnoreCase(OnboardingConstants.INVALID_PARAMETER)) {
				httpCode = HttpStatus.BAD_REQUEST;
			}
			return new ResponseEntity<ServiceResponse>(
					ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage(), modelName), httpCode);
		} catch (HttpClientErrorException e) {
			// Handling #401 and 400(BAD_REQUEST) is added as CDS throws 400 if apitoken is invalid.
			if (HttpStatus.UNAUTHORIZED == e.getStatusCode() || HttpStatus.BAD_REQUEST == e.getStatusCode()) {
				logger.debug(EELFLoggerDelegate.debugLogger,
						"Unauthorized User - Either Username/Password is invalid.");
				return new ResponseEntity<ServiceResponse>(
						ServiceResponse.errorResponse("" + HttpStatus.UNAUTHORIZED, "Unauthorized User", modelName),
						HttpStatus.UNAUTHORIZED);
			} else {
				logger.error(EELFLoggerDelegate.errorLogger, e.getMessage(), e);
				return new ResponseEntity<ServiceResponse>(
						ServiceResponse.errorResponse("" + e.getStatusCode(), e.getMessage(), modelName),
						e.getStatusCode());
			}
		} catch (Exception e) {
			MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
					OnboardingLogConstants.ResponseStatus.ERROR.name());
			MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, e.getMessage());
			logger.error(EELFLoggerDelegate.errorLogger, "onboardModel Failed Exception " + e.getMessage(), e);
			if (e instanceof AcumosServiceException) {
				return new ResponseEntity<ServiceResponse>(ServiceResponse
						.errorResponse(((AcumosServiceException) e).getErrorCode(), e.getMessage(), modelName),
						HttpStatus.INTERNAL_SERVER_ERROR);
			} else {
				return new ResponseEntity<ServiceResponse>(ServiceResponse
						.errorResponse(AcumosServiceException.ErrorCode.UNKNOWN.name(), e.getMessage(), modelName),
						HttpStatus.INTERNAL_SERVER_ERROR);
			}
		}

	}

	private Map<String, String> getArtifactsDetails() {
		List<MLPCodeNamePair> typeCodeList = cdmsClient.getCodeNamePairs(CodeNameType.ARTIFACT_TYPE);
		Map<String, String> artifactsDetails = new HashMap<>();
		if (!typeCodeList.isEmpty()) {
			for (MLPCodeNamePair codeNamePair : typeCodeList) {
				artifactsDetails.put(codeNamePair.getName(), codeNamePair.getCode());
			}
		}
		return artifactsDetails;
	}

	private String getArtifactTypeCode(String artifactTypeName) {
		String typeCode = artifactsDetails.get(artifactTypeName);
		return typeCode;
	}

	public String getCmnDataSvcEndPoinURL() {
		return cmnDataSvcEndPoinURL;
	}

	public String getCmnDataSvcUser() {
		return cmnDataSvcUser;
	}

	public String getCmnDataSvcPwd() {
		return cmnDataSvcPwd;
	}

}