# swagger-api-docs.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"Model upload server API.",
      "version":"1.0.0",
      "title":"Model Upload Server",
      "contact":{  

      }
   },
   "host":"localhost:8090",
   "basePath":"/",
   "tags":[  
      {  
         "name":"Onboarding Service APIs",
         "description":"Onboarding Controller"
      }
   ],
   "paths":{  
      "/v2/auth":{  
         "post":{  
            "tags":[  
               "Onboarding Service APIs"
            ],
            "summary":"Check User authentication and returns JWT token",
            "operationId":"OnboardingWithAuthenticationUsingPOST",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"cred",
                  "description":"cred",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/JsonRequest«Crediantials»"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "400":{  
                  "description":"Invalid request",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "500":{  
                  "description":"Something bad happened",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               }
            }
         }
      },
      "/v2/models":{  
         "post":{  
            "tags":[  
               "Onboarding Service APIs"
            ],
            "summary":"Upload model file and its meta data as string to dockerize",
            "operationId":"onboardModelUsingPOST",
            "consumes":[  
               "multipart/form-data"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "name":"model",
                  "in":"formData",
                  "description":"model",
                  "required":true,
                  "type":"file"
               },
               {  
                  "name":"metadata",
                  "in":"formData",
                  "description":"metadata",
                  "required":true,
                  "type":"file"
               },
               {  
                  "name":"schema",
                  "in":"formData",
                  "description":"schema",
                  "required":true,
                  "type":"file"
               },
               {  
                  "name":"Authorization",
                  "in":"header",
                  "description":"Authorization",
                  "required":false,
                  "type":"string"
               },
               {  
                  "name":"tracking_id",
                  "in":"header",
                  "description":"tracking_id",
                  "required":false,
                  "type":"string"
               },
               {  
                  "name":"provider",
                  "in":"header",
                  "description":"provider",
                  "required":false,
                  "type":"string"
               },
               {  
                  "name":"shareUserName",
                  "in":"header",
                  "description":"shareUserName",
                  "required":false,
                  "type":"string"
               },
               {  
                  "name":"modName",
                  "in":"header",
                  "description":"modName",
                  "required":false,
                  "type":"string"
               },
               {  
                  "name":"deployment_env",
                  "in":"header",
                  "description":"deployment_env",
                  "required":false,
                  "type":"integer",
                  "format":"int32"
               },
               {  
                  "name":"Request-ID",
                  "in":"header",
                  "description":"Request-ID",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "201":{  
                  "description":"Created",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "400":{  
                  "description":"Invalid request",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "401":{  
                  "description":"Unauthorized User",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               },
               "500":{  
                  "description":"Something bad happened",
                  "schema":{  
                     "$ref":"#/definitions/ServiceResponse"
                  }
               }
            }
         }
      }
   },
   "definitions":{  
      "Crediantials":{  
         "type":"object",
         "properties":{  
            "password":{  
               "type":"string"
            },
            "username":{  
               "type":"string"
            }
         }
      },
      "ServiceResponse":{  
         "type":"object",
         "properties":{  
            "errorCode":{  
               "type":"string"
            },
            "errorMessage":{  
               "type":"string"
            },
            "jwtToken":{  
               "type":"string"
            },
            "modelName":{  
               "type":"string"
            },
            "result":{  
               "type":"object"
            },
            "status":{  
               "type":"string"
            }
         }
      },
      "JsonRequest«Crediantials»":{  
         "type":"object",
         "properties":{  
            "request_body":{  
               "$ref":"#/definitions/Crediantials"
            },
            "request_from":{  
               "type":"string"
            },
            "request_id":{  
               "type":"string"
            }
         }
      }
   }
}